PLANET-2146 - Adjust navbar to fit a Language selector flag
This removes custom hard-coded class names for the nav items, that we
don't use anyway.
Adds a raw filter, which should be safe for item titles. in order to
render language flag image.
The current language won't be displayed at all, so in a bilingual
scenario we would only have one flag there.