PLANET-3864 Add Optimizer setting and anti-flicker snippet
Ref: https://jira.greenpeace.org/browse/PLANET-3864

This simplifies our GTM code. It sets the dataLayer variable regardless
of cookie settings, but only fires up GTM on consent (if GDPR setting is
enabled). This was done to allow Optimizer snippet to work, that needs
to follow dataLayer initialization. It also adds an Optimizer setting
for the NROs that don't use it.