PLANET-4609 Remove focal points fields and various
* Originally intended to fix multiple bugs, but ended up rewriting most
of it as that was simply less work then fixing these things in the form
the code was in. The main problem was that the code was copied from the
gallery block, and kept the different attributes to store the images,
instead of using nested attributes which gutenberg supports. That in
turn meant that the majority of lines of code where dealing with
splitting these attributes, searching for the element at the right index
and then combining them again, making reading and debugging the code a
huge headache. Also the it is better to change these attributes now the
block isn't used yet, as doing it when blocks exist in production would
also involve migrating steps.
* Fix messages not working.