PLANET-4603 Introduce campaign sidebar
* Register a sidebar with campaign settings.
* Add function for creating post-meta controls.
* Add function for loading options from a theme's json file.
* Use new meta field without underscore for campaign_page_template, as
the underscore automagically prevents API usage. In all usages fall back
to old field if the new doesn't exist yet.
* Add ColorPaletteControl that wraps around ColorPalette and functions
similarly to other input controls. This is currently missing in
wordpress, but should it be added there then this can go.
* Add css rule for the editor for the body font, with a css variable
defaulting to the native wordpress editor font.
* Hide the unpin button for the custom sidebar.
* Setup a subscriber to changes in the `campaign_body_font` post meta
field that updates that css variable to the picked choice.