PLANET-4741 Do not show tags/share buttons if title is hidden
The tags and share buttons were being shown under the top nav bar,
making them hardly visible and impossible to click.
Don't include these elements if the page title is hidden.
This is a temporary fix while we await the design of where the share
buttons should be if the page title is hidden.