PLANET-4827 WIP Archive as a post type
* Add custom post type for archive.
* Setup search for that post type.
* Make decay params configurable.
* Add settings for including archived content for logged in users only,
or publicly.
* Show the query time to logged in users.

* Convenience function for when a filter is used just to provide a
value, so that this value can be provided without having to pass a
function but instead just the filter name and the value.

* Specify which fields to return from ElasticSearch. Otherwise all
data, including post_content is loaded in memory even though it's
not used.

* Use ElasticSearch aggregations for tag and category data, so that we
can limit the amount of posts that are fetched to the a mount per page.

* Remove the filter for the main query. We should ensure that only one
query is used (by either using the main query for search, or by ensuring
the main query is not executed, but that can be taken up later.

* Fix decaying not being applied in ajax requests.
* Fix date ordering not working in ajax requests.
* Fix mime type filter not being applied in ajax requests.