PLANET-4838 Hide confusing WPML global settings on post edit page.
There is already a place for this config in the settings and there is
little value in having it at the bottom of each post edit page.
Furthermore these settings don't get saved by saving the post, instead
they have their own button which is small and not very noticeable in the
bottom right corner.