PLANET-4975 Fix document type filtering and page count
* Put the document type filter on all queries, by making it an OR query:
either the post type is something else than attachment, or the mime type
is one of the supported types. Remove previous filters that were in
multiple places as the new filter can be used on any result set.
* Don't show numbers lower than 0 for calculated page amount. We
subtract the Act pages from this as they have their own metric, however
in some cases the counts are not reliable and we might end up below 0.
Probably we want to replace this calculation with an aggregation that
already subtracts the Act pages.
* Add a link to edit page for logged in users.