PLANET-4923 Prevent the main query in search
* We actually don't use the main query and execute a query ourselves to
perform the search, so we can make the main query return an empty and
call the ElasticPress manually so that the main query doesn't hit the DB
or ES. This can't be done in master-theme, as that is only loaded after
the search is already performed.
* Wrapped in class exists as else phpunit tests fail for not finding
that class. Probably we should ensure that the unit test is setup with
the same plugins?