PLANET-5439 Class to facilitate parameterized queries
* Sql query parameters need to be be passed separately to $wpdb->prepare
and the values need to be in the same order as the placeholders in the
query. This makes the code hard to read and prone to bugs if a mistake
in the position is made. The class introduced here solves this problem
by allowing to specify these values in the query, where it will echo
back a placeholder and keep track of all values in the order they were
passed in. This ensure that the order will be the same, and that the
values can be seen in the position they will be used.