PLANET-5392 Redirect after features change.
* Some features affect what is shown in the WordPress admin menu.
However, because of how CMB2 works, these settings are only processed
after the menu admin menu has been rendered. This means that after
saving these features the admin menu does not correspond with these, a
page refresh is needed.
* Perform that page refresh in JS when there is a save success message
on the features page. The redirect cannot be performed on the backend,
because the header information is already sent before the settings are
processed.
* Modify the contents of the success message, so that it shows the page
is intentionally being reloaded.