PLANET-5292 Upgrade scripts + EN Form enable
* Add a way to run upgrade scripts in post deploy, in Activator. Fetches
a log of the scripts that were already run from the WP option, then runs
any new ones and adds these to the log. Finally persist the new log in
the option.
* Check if the corresponding script for the EN integration toggle has
already run. If it hasn't, the feature is considered as "on". This is
to get around a limitation: we can't give a boolean option a default
value of "on" with CMB2, as that would result in the option always being
on.