PLANET-5221 workaround as branch might not exist
* If you add a local path as a repository, composer will for some reason
only use it if a branch with the same name exists in git repository.
There could be a solution to make it also use the branches in the local
repository, but I didn't find it. If the branch does exist on github, it
will still use the local version because the commit sha was provided. So
a workaround is to create and checkout a fixed branch name after CircleCI
has checked out the code, and ensure that fixed branch name exists in
our repo. That branch in our repo can have any content, it just needs to
exist.