PLANET-4811 Support render from example props
* The block content is based on headings extracted from the current
post, which can't be done when rendering the example of the block, used
in the block's preview in the inserter, and the previews of the
different styles. To still display a preview with content this adds an
`exampleHeadings` attribute, which is only set in the example, along
with an `isExample` attribute (this could also have checked the presence
of `exampleHeadings`, but the boolean seems more clear.
* Wrap block register call in a function instead of a constructor.