PLANET-5640: Update Timeline lib to resolve google api version problems
Ref: https://jira.greenpeace.org/browse/PLANET-5640

The timeline block seems to be broken for some sheets urls.
The version 3.6.6 (introduced in https://github.com/greenpeace/planet4-plugin-gutenberg-blocks/pull/368)
calls a different url than 3.6.3 for the same document, and that url gets a 403 instead of the document itself.

Updating Timeline version to 3.8.10 solves this issue, and introduces 2 new ones:
- CDNJS doesn't have the lib, so we fetch it uncompressed (and unminified for CSS) from Knightlab CDN
- This version uses a proxy that we don't control for all spreadsheet calls