Hooks and hacks
* Turn VarPicker and VariableControl into function components that use
hooks.
* Add ability to unset a variable.
* Add refresh hack to make controls that read properties direclty
update.
* Fix a bug (?) in the ColorPicker component that made it not update
when its value prop changes. Maybe it's intended behavior that it
doesn't update. To force an update we use a ref and set the right
state on the component without triggering an input event on it. This
involved copying the function that formats the expected state, as
well as adding the tinycolor dependency of that function... Probably
there's a nicer way around this.