Remove preload and line break in href
* This line break caused an issue with Cloudflare's APO, which inlines
Google Fonts for better performance. They clearly didn't expect a
newline in there and did a bad request to Google resulting in a 400.
Cloudflare then proceeded to include that 400 response in the page, as
though it was the expected styles. This meant a html document was inside
of a style tag in the head, breaking the site very badly.
* Remove preload as APO doesn't remove it when inlining the google fonts
response.