PLANET-5733 "Fix" split2columns srcsets
* Fix typo in `issue_image_srcset`. Because of this type the attribute
was always removed, even though it was set correctly.
* Because the version attribute was never set (the default in JS block
definition is not saved), update_data runs on all blocks when viewing
the page on the front end. We can actually rely on this as PHP offers
ways to handle srcset that are not yet available in the editor.
* Remove srcset in the editor, since the fix for the front end won't
work there, but it's not a performance sensitive context.
* Use a const instead of a static variable.
* Set correct type on string attributes.
* Don't use 0 for ids that are not set, but null.
* Remove some casting to int on things that are already ints.
* Lengths can never be less than 0.

PLANET-5733 comply with counterproductive WP code style

* Align things.
* Remove return type of Base_Block::update_data, see comment.