Better color picker lib
* react-color has the same signature as WP's color picker, and works
better. Hence we don't need to update refs anymore.
* Reduce exported ACTIONS to just the keys with same keys as values. The
goal of putting the reducers in an object was to have IDE's point you
directly to the corresponding reducer when clicking the type in the
`dispatch` call. If this ACTIONS a separate constant with just strings,
this is much more tedious. My IDE still points me to the right place
even after only exporting the result of calling reduce on the reducers.