Remove browsable theme styleguide
Since we have Design System in place, we no longer require a browsable
styleguide, which can also be confusing. Removed all kss code and CI
jobs that rely on a url to run.

Also moved svg icons to master-theme.

Kept scss code in src because:
a. avoid changing app repos that use that
b. it actually makes sense