PLANET-6015 handle post locking in subscriber
* This was handled in a component's render method. So the post was only
accurately locked/unlocked when something triggered this component to
render. Somehow in WP 5.6 this was not working for us anymore, probably
because the component doesn't render as frequently anymore.
* Instead a subscriber is added here to update this on changes to the
post. HOWEVER...
* The subscriber runs very often, since Gutenberg stores everything in
one store. This might affect performance if we have heavy validation
logic. For now it seems ok since it only uses things from memory.
* To make things worse, the editor API allows access to post meta. But
this is not updated (sigh) when the CMB2 boxes are changed. Even now it
still relies on grabbing the metabox's value when you press the save
button, and doesn't update earlier like other things such as post
content or title. I got around this for the single metabox we need by
adding a listener that updates the store with the value on change. That
means the component can use `useSelect` to get updated when the meta
changes.

PLANET-6015 Add listener to update the meta store