Convert Cookies block to hooks
* Add new `useCookie` hook.
* Use constants for cookie name and values instead of string literals.
* Convert CookiesFrontend to hooks. It should have identical behavior as
the class did.
* Let `toAttribute` default to a empty function so that we don't need to
check its existence throughout the code. This should not be in the front
end component anyway, as this could result in unused dependencies on the
front end (not sure if it currently does). So instead of having 1
complex component that serves both front and back end, we could have a
much simpler component for both, with only minimal duplication.