Split webpack into multiple configs
* This is needed because we want to have separate aliases for public and
admin JS, so that we can use Preact for the former and WP's bundled
React for the latter.
* Also split off the CSS config, because we can. All config in the
previous mixed object was either applying to JS or to CSS, never to
both.
* Include blocks' entries one by one instead of using the template. This
wouldn't work for the split config, and there were already cases where
not all files were needed for a block. This was previously caught with a
file system check, which can make it hard to figure out what's
happening, and impossible to both have the file local and not build it.
* Remove the plugins that remove certain build artifacts. This was
probably added when the built files were checked into Git. Now that we
build these in CI this shouldn't be needed anymore. These are all small
files.