Save markup with ReactDOMServer.renderToString
* Otherwise `save` will use hooks in the context of the post editor,
which then causes invariant violations somehow. I suppose it's because
the registered `save` function of the block is not called at the top
level of a component by the block editor code. Maybe the block editor
should be using `renderToString` in the first place with the registered
component to avoid this issue?