Remove prop that only results in console warnings
* For some reason this was added to many RichText components, even
though they don't have this prop. So the only effect it had was cause a
whole lot of warnings flooding the console.
* This was copied from a component that was using the
`withCharacterCounter` HOC, that adds the counter and warnings around a
component. Don't use that component, HOCs are legacy, but we still need
to rewrite it. If we need to add a character counter in some places, we
should first rewrite the HOC.