PLANET-6307 Disable ButtonEdit override
Ref: https://jira.greenpeace.org/browse/PLANET-6307

---

Our overrides currently break buttons in WP 5.8. We can remove it for
now, and most of our overrides will still work by adding `theme.json`
on master-theme.

On the long term we should switch to `theme.json` for all native block
overrides but it currently doesn't support all the things we do for
buttons.