Set all attributes directly in listener
* The previous code resulted in the editor immediately having changes
just by opening it. That means that when you try to reload that page or
close it, you will get a blocking prompt, even if you didn't do any
action. For a user this is confusing because he has no idea what these
changes are, and it doesn't happen on every post.

Ref: https://jira.greenpeace.org/browse/PLANET-6436

Ensure entire tag data is not saved in attributes

* The attributes only need to have the ids, as the block is still back
end rendered. Unintentionally this made the block as big as the amount
of data is on a tag, which theoretically could be quite a lot, many
times bigger than the usual size of a post they're used in.
* Ensure not only the first 50 tags are fetched, we need them all.

Prevent modifying post without user interaction

* Unlike the code comment said, this is not needed at all. Because not
every single post is opened in the editor after we deploy these changes,
we have BC logic running anyway to catch those.