Add option to disable purge on feature change
Ref: https://jira.greenpeace.org/browse/PLANET-6256
* Purging all pages can send very large amounts of URLs to be purged,
and Cloudflare's purge API is designed for targeted purges. So while it
does work, it probably leads to us hitting a sort of rate limit, which
also affects production's ability to purge.
* On test instances we can leave this off because most of the time we're
logged in while testing. You can also always get a non-cached copy using
a hard refresh, or save a page to do a targeted purge.