PLANET-5934 Define 8px spacers
Ref: https://jira.greenpeace.org/browse/PLANET-5934

---

This approach removes all `$nxx` units (which are based on 5px
intervals anyway) and replaces `$space-breakpoint` variables
with a new naming scheme (check `_variables` file in this diff).

The new naming scheme uses a default `8px` spacer as
the base unit and then we create variations of that.

To make the variables easier to parse when reading the code,
it uses this pattern:
- `$sp-n`: means `n * 8px`
- `$sp-nx`: means `n.5 * 8px`

So for instance:
- `$sp-4` is `32px`
- `$sp-5x` is `44px`

---

Existing spacing values are replaced. Most of them could be
replaced with the exact same number. Those few that are not
already based on `8px` were by most `2px` off from their
nearest variable, which seems acceptable.

On old navbar and country selector just replaced with
numerical values since these are going to be deprecated soon.