Country Selector: Remove white space from button
Adjusted the markup to avoid a white space being rendered on both sides
of the text. It makes the code slightly less readable, but we avoid
having the underline decoration overflow on hover.

To compensate for the loss of the white space, added a bit more margin
to the two icons.