Use consistent attribute names for all Covers styles
This was causing console warnings for missing keys issues. It also makes more sense to use the same names everywhere as the style of Covers should only impact the CSS and not the data used. In case some existing blocks get the old data format because of cache issues when releasing, for now we kept fallback values to make sure nothing breaks.