PLANET-6603 Use the default fields from P4 settings in Cookies block
These fields were added to be used in the Cookies box with settings, but it makes sense to also use them as default values in the Cookies block. We also added reset buttons in case an editor has updated some of the Cookies block fields and wants to reverts them to the default values defined in the settings. This commit also includes adding a separate CSS file for Cookies block backend, for better for performance.