PLANET-6619 Redirect original page if it's a tag page override
* Since tag.php still includes page.php, it will again do this check.
But we can rely on WP's is_tag function to know it doesn't need to
redirect.

Use a permanent redirect so it gets cached

* Default is a 302.
* We need to check if this is safe to do, for example a redirect page
might be in a menu or otherwise linked to.

Use get_terms so that it caches the query

* This prevents the SQL query in most cases.
* It allows to already write it compatible with all taxonomy types, in
case we would use the mechanism in other places.
* Still need to check if the approach of caching per page makes sense.
Perhaps it's better to fetch and cache all tags redirect page data, so
that each page can read from the same cache.

Add a toggle so we can merge

* There's some things to sort for a few NROs that already redirect in
another way. By creating a toggle for this we're unblocked from merging.
* Show the toggle on production already.