Remove unnecessary CSS specificity
For some reason people still think it's a good idea to make CSS many
times more specific than it needs to be.

This is a major pain when refactoring code. Especially if the HTML also
contains tons of unnecessary wrapper elements which could be handled
much better by using recent CSS on the actual semantic markup.

Because the CSS mimics this structure, both become much harder to change
without breaking the other.

Instead, please try to write CSS that is reusable in multiple
circumstances. With the model of mimicing your DOM tree in CSS
selectors, you basically guarantee that each feature will need its own
collection of huge and specific CSS for declaring essentially the same
rules.