Make transition work also when closing
* The transition was only defined on the `.open` state, making it
disappear instantly when closing. It looks a lot better if it also does
the transition when closing.
The previous was added as a fix to prevent the menu showing on screen
resizes. This is addressed in another way here.

Hide mobile menu while closed

* And unhide it with no delay once it's opened.
* In order to allow the fade out effect to complete, add a delay of the
same duration.

Simplify animation CSS

* There doesn't seem to be a need to repeat styles with another value
than 100%, it all produced the same animation.
* By using a `:not(.open)` selector, we can avoid needing to specify
certain properties twice, such as visibility.

Use transform only for mobile menu slide in animation

* It's actually easier to use here, as 100% always refers to the element
width. So we don't need different CSS depending on how wide it is.
* Move it to a mixin as now it's only coupled to the .open class. This
helps keep the main class a bit readable as it's clearly all related to
the slide effect.