PLANET-6778 Add spacing to core Group block (#865)
* PLANET-6778 Add spacing to core Group block

Since we can use Group blocks within other blocks, this custom spacing should only apply to parent Group blocks. In this commit, we also add spacing between blocks in the editor (to be consistent with the frontend spacing) and add custom margins to numbered lists (to be the same as bullet lists).