PLANET-7020 Add new site identity feature toggle
This toggle is in Appearance > Customize > Site Identity. We also added this value to planet4_options, to be consistent in multilingual sites. This commit also includes the changes for PLANET-6984, to use the new grey color in body and headings when the new setting is toggled. This required a little bit of refactoring to work properly.