Add Setting for custom Tag Manager scripts URL
Adds an additional setting in the Analytics section of P4 settings for a custom domain name to load Tag Manager scripts from. The domain is then used in two places where the tag manager scripts are loaded. When the setting is left empty, the default domain of Tag Manager is used.