Remove redundant class check
This creates conflct issues with plugins that use a class with the same name
and not properly use PSR-4.

Example exception from a php library used in many plugins:

```
PHP Fatal error: Cannot declare class Action_Scheduler\Migration\Controller,
because the name is already in use in
action-scheduler/classes/migration/Controller.php on line 21
```

Removing the check, resolves the issue.