Simplify CF zone code to only remove www. from hostname
We made a change to accommodate for domains with more parts (like `.org.au`),
but that broke `ummah4earth.org` because they don't have `www` in their hostname.

Since we practically only want to remove `www`, this simplifies that part of the
code to avoid complicated conditional logic for all use cases.