Add translation text-domain for Blocks
As we moved Blocks to the theme, we also adjusted the Loco plugin
so that blocks translation also live here. They are already checked in
under `languages/blocks`.

This change just adds the additional translation text-domain, so the theme
know where to pick up the translations from.

Also deletes the backup autosave files. The script on the Handbook side has
been adjusted not to check in these files.