Register footer menus as theme navigation locations
Register footer menus so they are available as positions in WordPress admin.

Currently, the menus in the footer (Footer Primary, Footer Secondary, Social)
are only defined by naming convention. That doesn't work for multilingual
setups because menu names are unique and menus that work with the master
theme can not be created in more than one language at a time.

Registering the menu as theme locations enabled the use of WordPress admin
to select what menus to show in the footer, independent of menu names.

If no menu is set for the location, the old naming convention is still
in use to ensure backwards compatibility.

---

Ref: https://jira.greenpeace.org/browse/PLANET-7735