Prevent Akismet from firing before Turnstile
We use `preprocess_comment` with default priority 10 to fire Turnstile on comment submission.

But Akismet also uses the same filter with priority 1 (see `class.akismet.php` in plugin code).
So it actually runs before and in addition to Turnstile.

To avoid that, we can temporarily prevent Akismet from firing and then scan the comment
only when Turnstile challenge passes.